+++ src/test/java/edu/illinois/library/cantaloupe/resource/iiif/v1/InformationResourceTest.java
@@ -552,6 +552,8 @@ public class InformationResourceTest extends ResourceTest {
                 headers.getFirstValue("Content-Type")));
         // Date
         assertNotNull(headers.getFirstValue("Date"));
+        // Last-Modified (only present for cached infos, tested separately)
+        assertNull(headers.getFirstValue("Last-Modified"));
         // Link
         assertTrue(headers.getFirstValue("Link").contains("://"));
         // Server
@@ -568,6 +570,13 @@ public class InformationResourceTest extends ResourceTest {
         // X-Powered-By
         assertEquals(Application.getName() + "/" + Application.getVersion(),
                 headers.getFirstValue("X-Powered-By"));
+    }
+
+    @Test
+    void testGETLastModifiedResponseHeaderWhenDerivativeCacheIsEnabled()
+            throws Exception {
+        URI uri = getHTTPURI("/" + IMAGE + "/info.json");
+        tester.testLastModifiedHeaderWhenDerivativeCacheIsEnabled(uri);
     }
 
     @Test